aqme_DescriPyTor/requirements.txt
pandas==2.1.4
rdkit==2023.9.3
python-igraph==0.11.3
XlsxWriter==3.1.9
ipywidgets==8.1.1
pyarrow==14.0.2
plotly==5.18.0
customtkinter==5.2.1
chardet==5.2.0
matplotlib==3.8.2
rmsd==1.5.1
networkx==3.1
dash==2.14.2
pyvista==0.43.3
pyvistaqt==0.11.0
morfeus-ml==0.7.2
scikit-learn==1.3.2
seaborn==0.13.2
Pillow==10.3.0   # (instead of PIL)
scipy==1.11.4
tqdm==4.66.1
statsmodels==0.14.1
adjustText==1.2.0
multiprocess==0.70.17
shap==0.48.0
pymc==5.12.0
tabulate==0.9.0
ipykernel==6.28.0